Update copy_data.md
While in the process of getting more storage to store the correct files in the folder, should we link the other google drive that the CA ISRM is in? I think we would just need to add in the code below after the link, but it's totally optional! Also, let me know if you would like a screenshot for the part of the Test Checkup that says screenshot!

The CA ISRM data in the ISRM Tool Data Google Drive must be downloaded separately on [Google Drive](https://drive.google.com/drive/folders/1u9Y-bJmnYjlPrPje63q9QZDdwsA6nsGe).